+++ scripts/memory/tests.js
@@ -46,6 +46,30 @@ function makeRegistry(callback, reject) {
   return registry;
 }
 
+// This is not technically a memory-related test, but it depends on the build
+// artifacts generated in the ../../dist directory by `npm run build`, which is
+// an assumption shared by the other tests in this file.
+describe("@apollo/client/apollo-client.cjs", () => {
+  it("can be imported as a single CommonJS bundle (issue #8592)", () => {
+    const bundle = require("@apollo/client/apollo-client.cjs");
+
+    // Very basic test that requiring the bundle worked.
+    assert.strictEqual(typeof bundle.ApolloClient, "function");
+    assert.strictEqual(typeof bundle.InMemoryCache, "function");
+
+    // TODO This will change in AC4 when we move all React exports to the
+    // @apollo/client/react entry point (see issue #8190).
+    assert.strictEqual(typeof bundle.ApolloProvider, "function");
+
+    // The CommonJS bundles referred to by the "main" fields in the various
+    // package.json files that we generate during `npm run build` are all
+    // independent, non-overlapping bundles, but apollo-client.cjs is its own
+    // bundle, so importing it duplicates everything.
+    assert.notStrictEqual(bundle.ApolloClient, ApolloClient);
+    assert.notStrictEqual(bundle.InMemoryCache, InMemoryCache);
+  });
+});
+
 describe("garbage collection", () => {
   itAsync("should collect client.cache after client.stop()", (resolve, reject) => {
     const expectedKeys = new Set([